requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyMuPDF==1.23.8
Pillow==10.2.0
Werkzeug==3.0.1
stripe==7.9.0
python-dotenv==1.0.0
